requirements.txt
streamlit
openai
pinecone
pillow
numpy
requests
langchain
langchain-openai
langchain-community
langchain-pinecone
faiss-cpu
